+++ src/App.tsx
@@ -1,4 +1,4 @@
-import FetchData from './components/Ch1:Fetch_API/FetchCall';
+import FetchData from './components/Ch1-Fetch_API/FetchCall';
 
 export default function App() {
     return (
